Codewars Adding Big Numbers
The solution is correct, but the online compiler in Codewars does not recognize the "BigInt" function and does not accept the solution. Additional note: Sometimes, when the "BigInt" function is used in mathematical operations, it may perform incorrect operations. So try to use this function for small numbers. For larger numbers use "n". (23214832432432423423423n + 248324983294832432948n)